+++ tests/test_canvas.py
@@ -1,4 +1,6 @@
-from cs3560cli.lms.canvas import parse_url_for_course_id
+import requests
+
+from cs3560cli.lms.canvas import CanvasApi, parse_url_for_course_id
 
 
 def test_parse_url_for_course_id():
@@ -17,3 +19,81 @@ def test_parse_url_for_course_id():
         )
         is None
     )
+
+
+class MockSuccessfulGroupSetResponse:
+    @property
+    def status_code(self):
+        return 200
+
+    @staticmethod
+    def json():
+        return {
+            "data": {
+                "course": {
+                    "groupSetsConnection": {
+                        "nodes": [
+                            {
+                                "name": "Term Project Teams",
+                                "groupsConnection": {
+                                    "nodes": [
+                                        {
+                                            "name": "test-team-1",
+                                            "membersConnection": {
+                                                "nodes": [
+                                                    {
+                                                        "user": {
+                                                            "email": "rb000000@ohio.edu",
+                                                            "name": "Rufus Bobcat1",
+                                                        }
+                                                    },
+                                                    {
+                                                        "user": {
+                                                            "email": "rb000001@ohio.edu",
+                                                            "name": "Rufus Bobcat2",
+                                                        }
+                                                    },
+                                                ]
+                                            },
+                                        },
+                                        {
+                                            "name": "test-team-2",
+                                            "membersConnection": {
+                                                "nodes": [
+                                                    {
+                                                        "user": {
+                                                            "email": "rb000002@ohio.edu",
+                                                            "name": "Rufus Bobcat2",
+                                                        }
+                                                    },
+                                                    {
+                                                        "user": {
+                                                            "email": "rb000003@ohio.edu",
+                                                            "name": "Rufus Bobcat4",
+                                                        }
+                                                    },
+                                                ]
+                                            },
+                                        },
+                                    ]
+                                },
+                            }
+                        ]
+                    }
+                }
+            }
+        }
+
+
+def test_get_groups_by_groupset_name(monkeypatch):
+    def mock_post(*args, **kwargs):
+        return MockSuccessfulGroupSetResponse()
+
+    monkeypatch.setattr(requests, "post", mock_post)
+
+    client = CanvasApi(token="fake-token")
+
+    groups = client.get_groups_by_groupset_name("0", "Term Project Teams")
+    assert len(groups) == 2
+    groups = client.get_groups_by_groupset_name("0", "Homework 1")
+    assert groups is None